+++ app/src/main/kotlin/org/fossify/gallery/activities/SetWallpaperActivity.kt
@@ -1,11 +1,11 @@
 package org.fossify.gallery.activities
 
-import android.app.Activity
 import android.app.WallpaperManager
 import android.content.Intent
 import android.graphics.Bitmap
 import android.net.Uri
 import android.os.Bundle
+import androidx.core.graphics.scale
 import com.canhub.cropper.CropImageView
 import org.fossify.commons.dialogs.RadioGroupDialog
 import org.fossify.commons.extensions.checkAppSideloading
@@ -17,12 +17,14 @@ import org.fossify.commons.models.RadioItem
 import org.fossify.gallery.R
 import org.fossify.gallery.databinding.ActivitySetWallpaperBinding
 
-class SetWallpaperActivity : SimpleActivity(), CropImageView.OnCropImageCompleteListener {
-    private val RATIO_PORTRAIT = 0
-    private val RATIO_LANDSCAPE = 1
-    private val RATIO_SQUARE = 2
+class SetWallpaperActivity : BaseCropActivity() {
+    companion object {
+        private const val RATIO_PORTRAIT = 0
+        private const val RATIO_LANDSCAPE = 1
+        private const val RATIO_SQUARE = 2
+        private const val PICK_IMAGE = 1
+    }
 
-    private val PICK_IMAGE = 1
     private var aspectRatio = RATIO_PORTRAIT
     private var wallpaperFlag = -1
 
@@ -30,6 +32,9 @@ class SetWallpaperActivity : SimpleActivity(), CropImageView.OnCropImageComplete
     lateinit var wallpaperManager: WallpaperManager
 
     private val binding by viewBinding(ActivitySetWallpaperBinding::inflate)
+
+    override val cropImageView: CropImageView
+        get() = binding.cropImageView
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -62,7 +67,7 @@ class SetWallpaperActivity : SimpleActivity(), CropImageView.OnCropImageComplete
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, resultData: Intent?) {
         if (requestCode == PICK_IMAGE) {
-            if (resultCode == Activity.RESULT_OK && resultData != null) {
+            if (resultCode == RESULT_OK && resultData != null) {
                 handleImage(resultData)
             } else {
                 finish()
@@ -92,7 +97,6 @@ class SetWallpaperActivity : SimpleActivity(), CropImageView.OnCropImageComplete
 
         wallpaperManager = WallpaperManager.getInstance(applicationContext)
         binding.cropImageView.apply {
-            setOnCropImageCompleteListener(this@SetWallpaperActivity)
             setImageUriAsync(uri)
         }
 
@@ -137,33 +141,31 @@ class SetWallpaperActivity : SimpleActivity(), CropImageView.OnCropImageComplete
 
         RadioGroupDialog(this, items) {
             wallpaperFlag = it as Int
-            binding.cropImageView.croppedImageAsync()
+            cropImage()
         }
     }
 
-    override fun onCropImageComplete(view: CropImageView, result: CropImageView.CropResult) {
-        if (isDestroyed)
+    override fun onImageCropped(bitmap: Bitmap?, error: Exception?) {
+        if (isFinishing || isDestroyed) return
+        if (error != null || bitmap == null) {
+            toast("${getString(R.string.image_editing_failed)}: ${error?.message}")
             return
+        }
 
-        if (result.error == null && result.bitmap != null) {
-            toast(R.string.setting_wallpaper)
-            ensureBackgroundThread {
-                val bitmap = result.bitmap!!
-                val wantedHeight = wallpaperManager.desiredMinimumHeight
-                val ratio = wantedHeight / bitmap.height.toFloat()
-                val wantedWidth = (bitmap.width * ratio).toInt()
-                try {
-                    val scaledBitmap = Bitmap.createScaledBitmap(bitmap, wantedWidth, wantedHeight, true)
-                    wallpaperManager.setBitmap(scaledBitmap, null, true, wallpaperFlag)
-                    setResult(Activity.RESULT_OK)
-                } catch (e: OutOfMemoryError) {
-                    toast(org.fossify.commons.R.string.out_of_memory_error)
-                    setResult(Activity.RESULT_CANCELED)
-                }
-                finish()
+        toast(R.string.setting_wallpaper)
+        ensureBackgroundThread {
+            val wantedHeight = wallpaperManager.desiredMinimumHeight
+            val ratio = wantedHeight / bitmap.height.toFloat()
+            val wantedWidth = (bitmap.width * ratio).toInt()
+            try {
+                val scaledBitmap = bitmap.scale(wantedWidth, wantedHeight)
+                wallpaperManager.setBitmap(scaledBitmap, null, true, wallpaperFlag)
+                setResult(RESULT_OK)
+            } catch (ignored: OutOfMemoryError) {
+                toast(org.fossify.commons.R.string.out_of_memory_error)
+                setResult(RESULT_CANCELED)
             }
-        } else {
-            toast("${getString(R.string.image_editing_failed)}: ${result.error?.message}")
+            finish()
         }
     }
 }